added comment

+++ final/ai_cluegiver.py
@@ -175,7 +175,7 @@ def get_clue(words_obj, given_clues):
 		# top_scores = calculate_best_clue(all_possible_combos, good_words_obj_clues, bad_words_obj_clues, good_words_obj_dvf, bad_words_obj_dvf, stemmed_board_words)
 		# NEED TO ADD CODE TO CHECK IF THERE IS ONLY ONE WORD LEFT AND GIVE A CLUE FOR THAT ONE WORD
 		start_time = time.time()
-		calculate_best_clue_list_of_lists = [(all_possible_combos[i:i+4], good_words_obj_clues, bad_words_obj_clues, good_words_obj_dvf, bad_words_obj_dvf, stemmed_board_words, given_clues) for i in range(0, len(all_possible_combos), len(all_possible_combos)/20)]
+		calculate_best_clue_list_of_lists = [(all_possible_combos[i:i+4], good_words_obj_clues, bad_words_obj_clues, good_words_obj_dvf, bad_words_obj_dvf, stemmed_board_words, given_clues) for i in range(0, len(all_possible_combos), int(len(all_possible_combos)/20))]
 		with Pool() as pool:
 			try:
 				top_scores = pool.starmap(calculate_best_clue, calculate_best_clue_list_of_lists)
